chore(.gitignore): use github gitignore
Following the recommendation from Rick Reilly[1], add the gitignore file
from Github to the project. Mars[2] suggests using the GitHub for Unity
plugin, but also suggests Reilly's approach as an alternative. Simply
using a plugin can be a better alternative, but it is also one that will
avoid the trouble of going through all the pitfalls and errors that
result in learning. At this point it is more important to learn than to
succeed, hence the decision to go with the manual approach.

[1] REILLY, Rick. 2017 https://thoughtbot.com/blog/how-to-git-with-unity
[2] MARS, John. 2019 https://medium.com/@j0hnm4r5/git-with-unity-92ca7854ea3d