chore(.gitattributes): use Reilly's .gitattributes
Reilly[1] suggests using these entries for the gitattributes. The
entries contain file extensions for many popular files for models,
audio, images and fonts and Reilly claims that these are all the
filetypes for these sort of files Unity3D supported at the time.

Sentinels will rely on third party artwork that should include images,
audio, models and fonts and there is no guarantee that these will come
in a small number of different filetypes. Therefore it is wise to trust
Reilly's experience and go for a comprehensive .gitattributes file.