feat(assets): add StarSparrow assets
Adds the Star Sparrow modular spaceship asset from Ebal Studios. This
asset is available under the Standard Unity Asset Store EULA.

To the best of my knowledge I do not have the right to distribute the
asset separate from the game. I believe this is applicable to
distributing it in GitHub as versioned files. Therefore only the meta
file will be versioned here.