feat(utils): adds a utility to get subarrays from arrays
Apparently C# does not provide a convenient way to take subsections from
an array, but it does provides a very interesting feature called
extension methods. This feature allows one to expand the methods of
built-in classes with ease. There is a provided library called Linq that
is apparently very useful and leverages this feature to provide
query-like features to collections.

In this commit we add a SubArray features to objects that implement the
Enumerable interface.