feat(utils): symmetric apply
Provides a feature where one collection is applied a transformation to the first half and a second (potentially mirrored) transformation to the second half, leaving the last one untransformed.

The goal is to be able to transform the mirrored parts of ships with simple commands by ordering them in such a manner that one half is placed before the other half.