feat(tags): add ship part tags
Instead of identifying ship parts by name, we will use tags, allowing each ship part to have an arbitrary number of any given part.

It is possible that interesting ships can be built by having two bridges or two bodies, so we should not prevent that by requiring unique names for this. Tags seem to be a nice solution for this.