feat(GeneticAlgorithm): add matching func based on Leader's choice
Create a function to create matches that is based on the concept that the best performing individual chooses a partner.

The choice is basically the best performing individual among the others, disregarding it's own best performing achievement.

It is important to highlight here the difference between best performing achievement and best performing gene.

The fitness of each individual is computed based on a list of achievements. Each achievement does not have a direct correlation with any gene and for the purposes of the prototype, achievements are simply arbitrary.