feat(GeneticAlgorithm): add a GeneticAlgorithm factory
We now add a feature to be able to create as many GA pipelines as we wish, varying fitness function, selection function, matching function and crossover function.

Within the game this will allow us to have different evolution mechanisms that can contribute to the sensation that the game is not repeating itself.

Again, the GA in Sentinels will not have the goal of serving as the mechanism to enhance the difficulty as the game evolves, but to give a subjective perception that the player is facing something that adapts. Because of this, changing hyperparameters (even making them random) will generate different play experiences and the perception that something is adapting to the gameplay.