refactor(GeneticAlgoritm): adopt a procedural approach
Some uses of the Linq feature were hard to debug and on an effort to find the issues the code was converted to a procedural style. After the bug was fixed it was decided to leave the procedural code and not revert the changes as it was easier to work out the issues this way because there are intermediate variables we can use for debugging purposes.