refactor(ProtonLegacy): improve the ApplyGenome
The previous approach to ApplyGenome was not scalable. We have X number of parts and Y attributes for each part, therefore we need to create X*Y implementations of how is the gene to affect the part.

With this change we make it possible to write a standard transformation that will be applied to all parts and all attributes.

This approach will probably need to be improved for the final version, but it already allow us to see that it is more scalable.