feat(OrbitalFlight): first draft of the orbital flight
It provides a crude orbital flight. The ship will not look to the direction it is going and the speed variable controls the angle in which it is rotating.

During tests a small Earth was placed close to the ship and it successfully gave a feeling that it was a large Earth further away, but it seems this feeling will fade away when there are more objects in the scene.