doc: reset the README to describe the whole project
The first bit of code in this project was the Genetic Algorithm prototype. Now the project is evolving and new prototypes are being added. It is necessary to properly store the documentation for these new prototypes and also to describe the project as a whole.

We are re-using the repository from the first prototype so the README is re-written and the old one is moved to the Docs folder.

New prototypes will be created as separate branches and possibly separate scenes so that it is no longer necessary to have multiple repositories.