feat(maze generation): add sketch implementation of Wilson's algo
Wilson's algorithm for maze generations creates unbiased mazes.

The research on how to implement the stages led me to read about maze generation algorithms. Using a maze generation algo is a good strategy to generate fun stages.

Mazes are too complex for the ship to fly in, but a small maze can be a fun space station.

Wilson's algorithm is tweaked a bit because we are basically storing unity vectors. We are assuming that the class that will instantiate the space stations will create facilities in both sides of each "arrow", creating a path. The idea is also to make the maze more spare by adding random blanks in the maze walls.