fix: dependency issues.
After renaming Utils it is necessary to rename the Using statements.

Note: it was necessary to rename Utils because of an error. A Utils class had been created in the Maze Generation folder, but without the Maze Generation namespace. This file had not been commited, but as a result the code was importing the wrong Utils class. I didn't notice the problem until very recently.