fix(Maze Generation): out of bounds issue
We now validate the position before checking if it is in the visited board in order to avoid checking if potentially negative or too large positions are in the board.

Such too low and too large positions occur as we are walking through the path. Consider that current position is 0,0 and its direction is left or down, represented by -1,0 or 0,-1. The resulting position will be negative and will throw an error.

I such a situation we have finished adding the path to the visited maze.