feat(space station builder): instantiate prefabs and rotates
Preparatory code for Space Station Builder. It instantiates prefabs assuming all pieces are oriented towards the north and provide helper function to rotate the pieces based on a Position object describing a Direction.

A Direction is a position where both x and y are -1, 0 or 1, they are not the same and at least one of them is zero.

Requiring at least one to be zero makes the game play orthogonal, which is what we want for now.