feat(utils): grid and balanced range helper functions
This commit does 3 things. I'd like for it to be more specific, but I had trouble separating each part to commit individually.

First: rename SentinelsUtils to Utils. Previously I had renamed this class to SentinelsUtils to avoid a collision that I thought was going on between a supposed Unity3D Utils class and my own. There is no such Utils class in Unity3D, at least none that were colliding with mine. My error came about because I was misunderstanding the namespace and the using keywords in C#.

Second: We add a BalancedRange method to the Iterables section of Utils. This method returns a list of integer values with N numbers, starting with a negative number equal to floor(-N/2) and ending with floor(N/2). This methods will be useful to create positions around a center point without much further calculation.

Third, a CreateGrid method is created. It returns a pairwise combination of the values from two arrays. This is an interesting method as it required the use of Generics, which is a typing system that allows one to have strong types in situations where the type can only be known in the code that invokes the method, not in the declaration.