test(wilson algorithm): add more tests
There were something wrong with the code that the tests were catching only once in a while, probably something that only emerges depending on the value of some of the random values in the game.
We now add a hundred tests in a loop providing random values within a range to be able to catch the erros.