feat(SpaceStationBuilder): instantiage game parts from maze
SpaceStationBuilder.cs uses a Maze Generation Algorithm to decide where to create each part of a space station.

The maze is 2D array containing Directions. We use the directions basically to rotate the Game Objects and distribute the parts in a grid around the center of the SpaceStationBuilder GameObject.

The size of the grid is the size of the maze times _cellSize to allow the different part to avoid collision.