feat(Recycle Beam): first prototype of the recycle beam
This is an important feature for the game immersion.

The player should feel like controlling a ship that fights responsibly.
A weapon that recycles the enemies vessels parts is a relevant part of
this concept.

The sentinel is to have very few missiles and new ones are acquired by
recycling other ships. This also adds to the sentinels energy levels.

This commit adds the results of experimenting with the particle's system
to build a beam that looks like it is draining enemie's parts.

We create a particle emmiter that will emmit 3D cubes with a mesh
renderer using a material that denotes a piece of equipment. These
particles rotate and use a noise function to look like they were pulled
away from their origin.

This recycle beam will is to be used by positioning it onto the enemy
vessel and directing it to the player vessel.