feat: add collider to space station parts and add them to layer
Adding the space station prefabs to the SpaceStation Layer as a step for
the collision detection. Also adding a box collider to each of them.

The collider is a box collider for instead of a mesh collider to gain
some performance, given that we do not need a precise collision.