fix(Orbital Flight): make the ship look at its future position
After noticing a bug that resulted in the ship eventually flying
sideways, several approaches were attempted to fix it.

The bug came about because the ship would face a "Direction" (North,
South, West or East), meaning far away points that map to intuitive
directions. The problem is that in a circular orbit this would mean that
when flying further from the certer of the Earth the ship would twist.
This effect compounded resulting in weird flying. Also, when the ship
reverted its direction it would not revert the inclination.

The first attempt to fix this was to manually apply a rotation fix by
manipulating the Quartenion. This approach was unsuccessful.

A second approach was to replace the provided LooAt function with a
custom one that would handle this issue. This was also not successful.

A third approach, using Quartenion operations, consisted in creating a
Quartenion that was the inverse of the previous direction, apply it to
remove the direction from the Object rotation and then apply a new
quartenion built with the current direction. The idea was that the
object would be rotated by the Direction and then by the RotateAround function.
In the next frame we would like to preserve the effects of the
RotateAround function, but remove the effects of the Direction. So, an
Inverse of the Direction Quartenion was created and applied and then a
new Quartenion would be applied with only the new Direction.

This approach was successful in fixing the knwon issues, but the orbital
flight would still eventually result in a weird rotation due to the fact
that the game simulates a 2D top view flight over a sphere. As a result
a flight on the edges is faster than a flight in the middle, with
consequences to the ship rotation.

Finally, a new solution was to create a new GameObject (_aim) that would
be constantly poisitoned ahead of the ship using the same function
RotateAround. Regardless of the issues of applying a 2D plane over a
sphere, the ship would always face the direction it is going to.