feat(SpaceShip): add explosion behaviour
When a ship collides with a SpaceStation it explodes.
There may be more than 1 explosion, given that the collision is trigger.

This will allow the ship to fly throug a space station, leaving it in
flames. The ship may have a shield that will allow it to continue to
fly.