feat(MonteCarloTreeSearch): enemy ship navigation - WIP
This is a work in progress.

The research on creating the enemy ship behaviour had the following
requirements:

- we need the enemy ships to navigate between the space stations
- we need to be able to not feed each enemy ship with the space station
  actual positions. Providing a map would be acceptable.
- we need the algorithm to be fast and reasonably easy to implement.

While researching I found the paper bellow describing a goal oriented
navigation based on the Monte Carlo Tree Search algorithm.

After some sketches I arrived at a version of the algorithm that would
use child Game Objects as nodes and the Unity3D default collision
detection to assess collisions.
https://web.stanford.edu/class/aa228/reports/2019/final100.pdf

This implementation is based on a TreeNode interface that requires the
methods:
- GetChildren: returns the child TreeNodes
- IsBlocked: returns true if the node collides with a SpaceStation
- IsLeaf: returns true if the node has no children
- GetPosition: returns the position of the node

The algorithm will return the longest non blocked path. If there are
more than one such path, returns the closest to the target.