fix(MonteCarmeloTreeSearch): fix algo error
The previous implementation of the algorithm was yielding wrong results as the tests were showing.

It took a lot of time and many changes to reach this second implementation.

Here are some issues that were fixed:
- the new implementation uses new data structures, specially a Path data structure to ease the implementation
- Previous implementation had no mechanism to distinguish the different branches in the aggregation function. The new implementation relies of leafs and blocked nodes to create the Path object and then aggregates by choosing the the Path object and only transforms such object after the choice.

We can think of it like this.

Each node will choose the best path among it's children and add itself to it. Leaf nodes create Paths.