feat(WarpJump): Jumps to warp at the border of the board
This is an implementation of a suggestion from playtests.
The original goal was that the player would not be able to move far away from the space stations as the game would force the ship to come back.
In a playtest it was suggested that we should do just like Kobo Deluxe and make the ship appear on the other side and that perhaps we could add a warp jump so that the ship would come to the other side.