refactor(OrbitalFlight): make orbital flight a base class
There are several objects that perform orbital flight. The two most obvious are the Sentinel and the Enemy vessels. There are more, however, such as the camera and other floating objects.

This commit refactors the OrbitalFlight.cs in order to make it a base class. It is now a bit more abstract and two new classes were created: SentinelFlight.cs and MonteCarloFlight.cs. The latter should use the MonteCarlo algorithm to dodge spacestations when flying.