refactor(OrbitalFlight): both SentinelFlight and MonteCarloFlight inherit
OrbitalFlight now implements logic for setting the initial position and "follow me".
This was necessary to test the flight of the enemy vessels. It is necessary to have the camera follow the enemy vessel and to be able to set its position.

Setting the position will certainly be useful during the game also.