feat(Recycle Beam): pulls debris, distort view and throw ray
This is the hopefully final version of the main weapon visual effects.
This feature is important because it contributes to the immersion in the game.
The Sentinel does not attack by wasting energy or resources, but by grabbing matter and energy from the enemy vessel that is then used to boost it's own weapons.

In order to convey this notion and feeling we needed to make the ray pull debris from the enemy ship while at the same time making it clear that the objects that come torwards the player's vessel are not a threat.

The effects chosen were taken from several FX packages and customized. The floating debris were built from Unity3D's default resources.