feat(SimpleFlight): allow enemies to use a simple algorithm to guide their flight
A lot of time was invested into MonteCarloFlight.cs but it is still not working as expected. With this commit we move on and implement a very simple algorithm to replace that.

If there is time before the end of the project I may come back to this to fix the bugs.