fix: explosion and damage
The initial explosion object chosen would not render in WebGL. A new and simpler explosion object was chosen and used here. We also fixed the usage of the particle system by avoiding it to be destroyed when the execution ended.
This was implemented in response to a script that came with the new explosion. It would automatically destroy the particle system upon finishing. This is a feature that is also provided by the Editor.
When I finally noticed that this was being caused by the alien script and removed it these safe guards were already in place. Given that they do play a role in interrupting and restarting the FX, I'll leave it in the code.