refactor(Effects): create prefab/effects folder
As the project evolves it becomes clear that a better organization is needed. External prefabs are used from their original folders and become hard to find. Also, it is hard to clean up the project of unused external prefabs.
Another issue is that the separation of the Scripts folder and the prefab folder feels incompatible with the object oriented nature of the Unity framework. Most prefabs require a behaviour aspect that is implemented in a script. Having this separated might be making things more confusing. However, it is hard to be sure of this given that most of the confusion probably comes from the number of external assets folders.