feat(Moon): add a Moon object
The immersion experience can be enhanced by adding this familiar element to the game.
One of the questions for this week's test was if the game was too unrealistic.
There are several elements that completely fail in being realistic. One of the main issues is proportions.
Earth is very small and close to the player. This would hopefully not be easy to see given the lack of references. The absence of any other celestial body was another issue that this commit intends to solve.