refactor: reorganize folder structure
All game assets and scripts were living in the Assets folder, regardless of being imported as external assets or developed within the Sentinels project. Because of this it was harder than necessary to navigate the directory and find the assets needed.

I want to move into a practice where all used assets are moved into game specific folders so that the imported assets are removed afterwards.

The meta files should be commited anyway to keep a permanent record of the external assets used.