fix(Moon): increase Moon speed.
We are not looking to make this game realistic, but to make it
immersive. It would be nice if the players would notice the moon orbit
(even if it is not realistic) as that may cause a feeling of being
immersed in a real thing. It would certainly be nicer to make it
realistic but that would not only cause mucho more work but it could
also make it harder to experience as objects could become too far away
from each other.