fix(ExplosionSmoke): smoke vfx is displayed
SmokeVFX was previously a child of the ExplosionDestroyVFX and when the
latter was started the former would not automatically start. Now these
are separated. AudioSource is also moved from CoreParts to the root of
the ship and CoreParts are set inactive upon explosion.

As a result, when an enemy is destroied it an explosion and sound can be
heard but the ship is no longer perceived.