refactor(tests): split tests into editor and play mode
When first learning about unit tests I skiped implementing editor tests
as I initially thought those were destined to test addons to the editor
itselsf, which I would not do. As I gained experienced and learned that
was not the case I also noted that the time it takes to run play mode
tests has an important impact on my will to write and constantly run
tests.

There are bugs in the game currently that shouldn't be there should I
have been more judicious with test writing, so I'm trying to fix this in
the latest stage of the project.