fix(ProtonLegacy prefabs): enemy ships never take off
Upon creation the enemy ships were flying together with the headquarters and never departing. It's behaviour should greatly increase it's speed when it is away from the board, which was not happening.

It turns out that this was happening because the warp multiplier for the enemy ships was set to zero.