feat(ProtonLegacy): ProtonLegacy.cs is responsible for features
Converting from genome code into behaviour is done in two steps. ShipGenome specifies which genes are responsible for what, but it is ProtonLegacy that is responsible for mapping the encoded values into useful values. This architecture allows us to have at some point different families of enemies with different mappings.