feat: end game scenes
Add a scene for game over and for congratulating the player upon victory.

The congratulating screen shows the same image from the menu: a beautiful natural landscape. The goal is to convey a peaceful and balanced Earth that the player helped preserve.

The game over scene shows a catastrophe scenario.

Both scenes were created to adjust to different screen sizes. This feature was not necessary for this project but it was the way I found to make sure the image would be shown appropriately.