fix: use Update for user input & avoid diagonal boost
Using FixedUpdate for user input is not recomended.

It was tricky to get the game to work as smoothly with update as it was
working with fixedupdate, but it was also necessary. With fixed update
one off inputs would sometimes be completely missed, and this does not
happen with update.