fix(GA): apply genome to core parts
Ever since the core was created for the enemy vessel the mutations were not being reflected in the physical aspect of the ships.

That was because I was trying to change that appearence in the main gameobject instead of looking for the parts in the core.

The core was necessary so that we could have parts of the ship that would not affect its size and position.