feat(Recycle beam): add alternative key
From user feedback: using Ctrl as the key for firing is a poor choice as it makes the game unplayable in MacOS.

This is quite an interesting remark given that the left Ctrl is the default key for Fire1 axis in Unity3D.

With this change the "z" key will have the same effect. The key was chosen because it is placed close to the left ctrl.