fix(SpaceStationBuilder): board now contains all stations
This is from a bug report from the playtests. The board would end before the last station in the borders. That is because the game board ended when the y axis was as low as the most northern point in the board, which is not the farthest point.

This fix uses pythagorean theorem to find the furthest point and adds a padding to it.